Revert "KAN-29 Qn-service integration with frontend"
