+++ backend/api/.env.example
@@ -1,0 +1,12 @@
+# Authentication (Supabase JWKS)
+SUPABASE_JWT_URL=https://YOUR_PROJECT.supabase.co/auth/v1/certs
+SUPABASE_JWT_AUD=authenticated
+
+# MongoDB
+MONGODB_URI=mongodb+srv://user:pass@cluster.mongodb.net/?retryWrites=true&w=majority
+MONGODB_NAME=QuestionService
+MONGODB_COLLECTION=questions
+
+# Server
+PORT=4000
+
