+++ backend/api/src/auth/bearer-auth.guard.ts
