+++ backend/api/src/mongodb/mongo.provider.ts
@@ -54,13 +54,9 @@ const collectionProvider: Provider = {
     useFactory: async (db: Db): Promise<Collection> => {
         if (cachedCollection) {
             return cachedCollection;
-        }
+        } 
 
-        // Support both MONGODB_COLLECTION and legacy QUESTIONS_COLLECTION_NAME
-        const collectionName =
-          process.env.MONGODB_COLLECTION || process.env.QUESTIONS_COLLECTION_NAME || 'questions';
-
-        cachedCollection = db.collection(collectionName);
+        cachedCollection = db.collection(process.env.MONGODB_COLLECTION || 'questions');
         return cachedCollection;
     },
     inject: [MONGO_DB],
